Rename dotnet-roslyn to dotnet-roslyn-tools
