+++ RoslynTools.sln
@@ -66,9 +66,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GitHubCreateMergePRs", "src
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NugetDependencyFinder", "src\NugetDependencyFinder\NugetDependencyFinder.csproj", "{7A6963D6-1F40-4CC6-9941-7F912252F877}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "dotnet-roslyn", "dotnet-roslyn", "{C68033E0-FB8F-4327-BFB5-B340A5A7778F}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "dotnet-roslyn-tools", "dotnet-roslyn-Tools", "{C68033E0-FB8F-4327-BFB5-B340A5A7778F}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Roslyn", "src\dotnet-roslyn\Microsoft.Roslyn.csproj", "{1D14C80C-B4C3-42C1-BA4C-8A81427CA44D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.RoslynTools", "src\dotnet-roslyn-tools\Microsoft.RoslynTools.csproj", "{1D14C80C-B4C3-42C1-BA4C-8A81427CA44D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
