+++ src/dotnet-roslyn-tools/Program.cs
@@ -1,0 +1,4 @@
+using Microsoft.RoslynTools.Commands;
+using System.CommandLine;
+
+return await RootRoslynCommand.GetRootCommand().InvokeAsync(args);
