--- src/dotnet-roslyn/Logging/NullScope.cs
@@ -1,18 +1,0 @@
-// Licensed to the.NET Foundation under one or more agreements.
-// The.NET Foundation licenses this file to you under the MIT license.
-// See the License.txt file in the project root for more information.
-
-namespace Microsoft.Roslyn.Tool.Logging;
-
-internal class NullScope : IDisposable
-{
-    public static NullScope Instance { get; } = new NullScope();
-
-    private NullScope()
-    {
-    }
-
-    public void Dispose()
-    {
-    }
-}
