--- src/dotnet-roslyn/Microsoft.Roslyn.csproj
@@ -1,38 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-
-    <ToolCommandName>roslyn</ToolCommandName>
-    <Description>Command line tool for performing Roslyn infrastructure tasks.</Description>
-
-    <IsPackable>true</IsPackable>
-    <PackAsTool>true</PackAsTool>
-
-    <!--
-      These identifiers are for generating the shim'd core executables for signing.
-    -->
-    <PackAsToolShimRuntimeIdentifiers Condition=" '$(DotnetBuildFromSource)' != 'true' ">win-x64;win-x86;osx-x64</PackAsToolShimRuntimeIdentifiers>
-    
-    <!-- enable building symbol packages -->
-    <DebugType>portable</DebugType>
-    <IncludeSymbols>true</IncludeSymbols>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="System.CommandLine" Version="2.0.0-beta2.21615.1" />
-    <PackageReference Include="System.CommandLine.Rendering" Version="0.4.0-alpha.21615.1" />
-    <PackageReference Include="Microsoft.Extensions.Logging" Version="6.0.0" />
-    
-    <PackageReference Include="Azure.Identity" Version="1.6.0-beta.1" />
-    <PackageReference Include="Azure.Security.KeyVault.Secrets" Version="4.3.0-beta.4" />
-    <PackageReference Include="LibGit2Sharp" Version="0.26.2" />
-    <PackageReference Include="Microsoft.TeamFoundationServer.Client" Version="16.201.0-preview" />
-    <PackageReference Include="Microsoft.VisualStudio.Services.Client" Version="16.201.0-preview" />
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
-  </ItemGroup>
-
-</Project>
