--- src/dotnet-roslyn/Program.cs
@@ -1,4 +1,0 @@
-using Microsoft.Roslyn.Tool.Commands;
-using System.CommandLine;
-
-return await RootRoslynCommand.GetRootCommand().InvokeAsync(args);
