--- src/dotnet-roslyn/VS/IProduct.cs
@@ -1,18 +1,0 @@
-// Licensed to the.NET Foundation under one or more agreements.
-// The.NET Foundation licenses this file to you under the MIT license.
-// See the License.txt file in the project root for more information.
-
-namespace Microsoft.Roslyn.VS;
-
-internal interface IProduct
-{
-    string Name { get; }
-
-    string RepoBaseUrl { get; }
-    string ComponentJsonFileName { get; }
-    string ComponentName { get; }
-    string BuildPipelineName { get; }
-    string? PackageName { get; }
-    string? ArtifactsFolderName { get; }
-    string[] ArtifactsSubFolderNames { get; }
-}
